The whole script is devided into 3 part: main program to run, processing to process data and visualizaion for plotting the data.
