+++ .gitignore
@@ -1,3 +1,3 @@
 *.txt
 *.png
-
+__pycache__/
